+++ Tests/Public/Record/PrimaryKeyHiddenRowIDTests.swift
@@ -659,7 +659,7 @@ class PrimaryKeyHiddenRowIDTests : GRDBTestCase {
             let expectation = self.expectation(description: "expectation")
             let controller: FetchedRecordsController<Person>
             #if os(iOS)
-                controller = FetchedRecordsController<Person>(dbQueue, request: Person.all(), compareRecordsByPrimaryKey: true)
+                controller = try FetchedRecordsController<Person>(dbQueue, request: Person.all(), compareRecordsByPrimaryKey: true)
                 var update = false
                 controller.trackChanges(
                     tableViewEvent: { (_, _, event) in
@@ -674,7 +674,7 @@ class PrimaryKeyHiddenRowIDTests : GRDBTestCase {
                         expectation.fulfill()
                 })
             #else
-                controller = FetchedRecordsController<Person>(dbQueue, request: Person.all())
+                controller = try FetchedRecordsController<Person>(dbQueue, request: Person.all())
                 controller.trackChanges { _ in
                     expectation.fulfill()
                 }